Add files via upload
One night, I was looking for a stop watch on the internet, and I could not find a good one. But then, I start to make one. It is not beautiful, not all fantasy and colorful, but A simple stop watch with tree buttons; start, reset, and pause... and can count the time till 99 hours.